Add files via upload
Updated Arduino files which allow full control from the scope user panel.  New functions are to set the RTC date and time.  Also a new function allws you to set the file name stub in the form of XXXXXnnn.SVCG.  The XXXXX is a 5 character file name stub which is then followed by an automatically generated serialization number 000-999 so you do not have to constantly change the file name manually.